progress made on healer profile feature

+++ src/main/kotlin/com/kgoro/sangoma_link/address/AddressRepository.kt
@@ -2,5 +2,5 @@ package com.kgoro.sangoma_link.address
 
 import org.springframework.data.jpa.repository.JpaRepository
 
-interface AddressRepository:  JpaRepository<Address, Long> {
+interface AddressRepository:  JpaRepository<Address, Int> {
 }
